refactor: extract AstTypeResolutionHelpers and AstMethodLookupHelpers from AstUtilities
Extract 17 methods (8 type-resolution, 9 method-lookup) into two focused
helper classes, reducing AstUtilities.java from 649 to 475 lines.

AstTypeResolutionHelpers: property getter discovery, setter resolution,
keyword factory types, parameter value types, declaring-type resolution.

AstMethodLookupHelpers: method lookup by signature, SAM resolution,
overridden method discovery, invocation crawling, hierarchy traversal.

All 20 caller files updated (2 same-package, 18 cross-package).
createUserLambda now delegates to AstMethodLookupHelpers.getSingleAbstractMethod.
No behavioral changes — pure mechanical extraction with preserved signatures.

Closes #642
