+++ netbeans/src/test/java/org/alice/netbeans/project/ProjectCodeGeneratorGeneratedSourceTest.java
@@ -4,6 +4,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.lgna.project.Project;
+import org.lgna.project.ast.AstMethodLookupHelpers;
 import org.lgna.project.ast.AstUtilities;
 import org.lgna.project.ast.BlockStatement;
 import org.lgna.project.ast.BooleanExpressionBodyPair;
@@ -492,7 +493,7 @@ public class ProjectCodeGeneratorGeneratedSourceTest {
 
   private static NamedUserType programTypeWithForEachIterableMethod() {
     NamedUserType type = programType("Program");
-    JavaMethod asList = AstUtilities.lookupMethod(Arrays.class, "asList", Object[].class);
+    JavaMethod asList = AstMethodLookupHelpers.lookupMethod(Arrays.class, "asList", Object[].class);
     MethodInvocation iterable = new MethodInvocation(
         new TypeExpression(asList.getDeclaringType()),
         asList,
