+++ netbeans/src/test/java/org/alice/netbeans/project/ProjectCodeGeneratorStandaloneProjectTest.java
@@ -5,6 +5,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.lgna.project.Project;
+import org.lgna.project.ast.AstMethodLookupHelpers;
 import org.lgna.project.ast.AstUtilities;
 import org.lgna.project.ast.BlockStatement;
 import org.lgna.project.ast.JavaMethod;
@@ -531,7 +532,7 @@ public class ProjectCodeGeneratorStandaloneProjectTest {
 
   private static UserMethod mainMethodWithProbe() {
     UserParameter argsParameter = new UserParameter("args", String[].class);
-    JavaMethod recorder = AstUtilities.lookupMethod(
+    JavaMethod recorder = AstMethodLookupHelpers.lookupMethod(
         ProjectCodeGeneratorStandaloneProjectTest.class,
         "recordGeneratedProgramMainArgs",
         String[].class);
